requirements.txt
APScheduler==3.9.1.post1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
slack-bolt==1.15.5
slack-sdk==3.19.4
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
